feat(l1): only track state root of big accounts (#4262)
**Motivation**

When inserting storages ranges during snap sync, we keep track of the
storage root to insert to the same tree as we go from one chunk to
another. We don't need to keep track if the state root is already
complete.

**Description**

This PR only updates maybe_big_account_storage_state_roots if the
account is, in fact, big (couldn't be fetched in a single batch).