fix(l1): make account and storage ranges handlers async (#4401)
**Motivation**

These handlers are currently sync and can block the runtime for quite a
long time. This is not the ideal solution for this; we still need to do
things better so the handler itself doesn't take a long time to fetch
the requested nodes. We'll tackle this on a subsequent PR.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>