perf(l1): implement fast skip for TrieIter (#4404)
**Motivation**

There's been stalls in sync'd nodes due to receiving requests for
`GetAccountRanges` and `GetStorageRanges`.
The cause was that we iterate the whole state/storage trie from the root
to the leaves, manually skipping the irrelevant ones, which takes
minutes.
At the time, the operation was also blocking, freezing the runtime.

**Description**

Adds an `advance` method to avoid having to traverse the whole state
trie when requested account and storage ranges.
Adds methods to `Store` receiving the initial keys.
Uses those methods for snap capability handlers.