refactor(replay): remove `ProverDb` and use `ExecutionWitness` (#4440)
**Motivation**

This pull request refactors the caching and witness-handling logic in
the `ethrex_replay` tool to simplify and unify the use of execution
witnesses. The changes remove the previous abstraction of
`PreExecutionState` and instead use a single `ExecutionWitness`
structure for caching, running transactions, and generating program
inputs.

**Description**

* The `Cache` struct and related logic now store and use an
`ExecutionWitness` instead of a `PreExecutionState`, removing the need
for dual handling of database and witness states.
* The `RpcDB` struct now provides a `to_execution_witness` method that
builds an `ExecutionWitness`, replacing the previous `to_prover_db`
method.
* The `Trie::from_nodes` method now always requires a root hash and no
longer accepts a root node RLP.
* Removed all code paths and logic related to the
`PreExecutionState::DB` variant as the code now exclusively uses the
`ExecutionWitness`.