fix(l1): truncate file contents when printing DumpError (#4480)
**Motivation**

Whenever an error causes `DumpError` to be printed, the whole contents
of the file that caused the errors may be shown on screen, which can
cause crashes due to the sheer amount of data printed.

**Description**

This PR changes `DumpError`'s `Debug` implementation to one that
displays the length of `contents` instead of the whole contents of the
file.