fix(l2): race condition tracking privileged tx count (#4394)
**Motivation**

We have a race condition between the `block_producer` and the
`l1_committer` when updating the `precommit_privileged` table. The race
condition manifests as the following error:
```
2025-09-08T18:17:00.931203Z ERROR ethrex_l2::sequencer::l1_committer: L1 Committer Error: Committer failed to send transaction: Failed to send commitment for batch 3. first_block: 14 last_block: 38: Committer failed because of an EthClient error: eth_estimateGas request error: execution reverted: unknown error
```

This is what happens:

The `block_producer` tracks the number of privileged transactions in the
`rollup_store` to ensure it does not exceed the maximum cap of 300,
which is imposed by gas limitations. Once the cap is reached, it stops
adding privileged transactions until the `l1_committer` resets the
counter after creating a new batch.

But given the `block_producer` and `l1_committer`run concurrently, the
following scenario may occur:
1. The `l1_committer` starts producing a new batch containing the last
available block.
2. The `block_producer` creates a new block with 150 privileged
transactions and sets the counter to 150.
3. The `l1_committer` finishes the batch and resets the counter to 0.
4. The `block_producer` continues adding privileged transactions until
the cap is reached again.
5. As a result, the `l1_committer` builds a batch containing 450
privileged transactions instead of 300.

**Description**
- Removes the shared table in the `rollup_store`.
- The `block_producer` only enforces a `PRIVILEGED_TX_BUDGET` per block.
It no longer has a notion of batches.
- The `l1_committer` ensures that the `PRIVILEGED_TX_BUDGET` is not
exceeded per batch when adding blocks.
- The `block_producer` tracks the `last_privileged_nonce` to ensure all
privileged transactions are added sequentially across blocks. IMO, this
is a better solution since it allows us to enforce ordering across
different batches. However, it is not the ultimate solution, because we
still mark the very first processed privileged transaction as the first
one, even though it may not actually be in order (this issue already
exists and should be addressed in a separate PR). #4398 was created to
solve this.


Closes #4359