fix(l1): handle syncing in case of reorgs and handle the case of reorg with invalid side chain (#4028)
**Motivation**

Restoring `request_block_headers` previous to this
[pr](https://github.com/lambdaclass/ethrex/pull/4213) since it brought
issues in case of re org.
Fixing the Invalid Missing Ancestor Syncing ReOrG hive tests which are
failing.

**Description**

Up to now, when executing blocks during syncing, if there's an error
while running a batch we set only the failing block with having an
invalid ancestor. This pr sets all the downloaded descendants of a
failing block with having an invalid ancestor during syncing.

Also, when syncing, it drops the sync head header from the requested
headers when downloading block bodies if the sync head is stored as a
pending block, since we already have it and there's no need to ask for
it.

On another note, it was necessary to restore for full syncing the
`request_block_headers` that we had previous to merging this
[pr](https://github.com/lambdaclass/ethrex/pull/4213), since there was
an issue with the way the block headers were requested. The core problem
with the update was that we requested the all headers left from the sync
head on by number, and since the tests come with a re-org scenario and
the sync head number was the same as the current head number, no headers
were requested, so the fork choice update never stopped syncing.

Closes #3105