feat(l2): start/pause sequencer admin endpoint (#4367)
**Motivation**

We want to have more control over the sequencer, for this reason this PR
implements a new admin server with an API to start/stop the sequencer
commits. In the future this server can be expanded to provide a /health
endpoint with information about the sequencer status.

**Description**

- This PR adds a new admin http server that starts with the sequencer
and has access to the L1Committer mailbox to send start/stop messages
- Add to the L1Committer logic to handle a call message that stops the
sequencer
- Add to the L1Committer logic to handle a call message that start the
sequencer with an optional parameter of a start delay in ms
- Add 3 endpoints to the admin server they can be tested by running
`make init-l2-dev` and then using any of the commands:
- `/committer/start`: starts the committer `curl -i
localhost:5555/committer/start`
- `/committer/start/{delay}`: starts the committer with a delay in ms
`curl -i localhost:5555/committer/start/60000`
- `/committer/stop`: stops the committer `curl -i
localhost:5555/committer/stop`
- Add new page in docs for the admin server API