test(l1): added error line to get more detail on InvalidArgument error (#4501)
**Motivation**

A `MessageSendFailure(IoError(Os { code: 22, kind: InvalidInput,
message: "Invalid argument" }))` is spamming from time to time when
trying to send a discv4 message.

**Description**

Added an error line to detail the message being send, the destination
address and the internal error to get better insight of the problem

Once the problem is found, we can revert this code.

Closes #4492