fix(l2): avoid blocking in `L1Committer` (#4164)
**Motivation**

The L1 Committer does some blocking computations when publishing batches
to L1, which took from 100ms to 30ms according to quick measurements on
an M3. This is a low number, but still degrades the tokio runtime and
could cause issues elsewhere in the system.

**Description**

~This PR wraps the call to `generate_blobs_bundle` with a
`spawn_blocking`, to avoid blocking the tokio runtime. Also adds a
comment on another call that seems to also be blocking, taking sometimes
\~40ms before reaching an `await` point.~ Changes reverted. Now we
instead spawn the whole GenServer as blocking.