perf(l1): save code hash during account insertion and state healing (#4487)
**Motivation**

Improve bytecode download time saving code hash in files

**Description**

When we are doing snapsync, we see that we waste too much time
downloading bytecode since we are traversing the trie looking for the
code hash. Before this step, we are inserting the account state to the
trie so we could take advantage of this and start saving the code hash
in a buffer. When the buffer exceeds the chunk size, flush it to a file.

We have some constant values ​​that we can adjust and see the numbers.
- `BYTECODE_CHUNK_SIZE`
- `CODE_HASH_WRITE_BUFFER_SIZE`

Hoodi
- `main` - `bytecodes progress: downloaded: 719368, elapsed: 00h 03m 28s
00ms`
- `pr` - `bytecodes progress: downloaded: 710598, elapsed: 00h 00m 32s
00ms`

Sepolia
- `main` - `bytecodes progress: downloaded: 11090586, elapsed: 01h 43m
58s 06ms`
- `pr` - `bytecodes progress: downloaded: 7979995, elapsed: 00h 07m 59s
00ms`

Mainnet
- `main` - `bytecodes progress: downloaded: 4930763, elapsed: 03h 15m
14s 11ms`
- `pr` - `bytecodes progress: downloaded: 1813931, elapsed: 00h 02m 50s
00ms`
---

Closes #4474

---------

Co-authored-by: fedacking <francisco.gauna@lambdaclass.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>