feat(replay): replay use add block instead of stateless execution (#4396)
**Motivation**

- Use ethrex replay with `add_block` so we can measure the exact same
thing that we measure in normal runs.

**Description**

If we want a samply of `add_block` we can do this:
`cargo build --profile release-with-debug -p ethrex-replay`
`samply record target/release-with-debug/ethrex-replay block --rpc-url
'http://65.109.97.102:8545' --execute --no-zkvm`

If searching for `add_block` in samply we will see something like this:
<img width="1869" height="936" alt="image"
src="https://github.com/user-attachments/assets/dd128dc2-422c-4c04-bf9c-8eded8838621"
/>


We had to add feature flag replay so that this functionality works with
reth nodes. Otherwise it only works with ethrex nodes that behave the
same as us and therefore require the exact same trie nodes and codes.