fix(l1): fixed edge case for missing accounts in state trie in storage healing (#4606)
**Motivation**

The storage healing algorithm was assuming that missing accounts in the
state trie were a bug. Due to an edge case some accounts can be deleted
from state trie. As such, the code now filters healed accounts that
don't exist in the trie.

**Description**

- Changed storage healing `get_initial_downloads` to filter healed
accounts that aren't present in the trie.
- Changed p2p logging to include header hash