feat(l1): handle non-canonical headers by hash in `GetBlockHeaders` (#4576)
**Motivation**

We're currently returning empty header responses when we receive a
`GetBlockHeaders` request specifying the start with the block's hash,
and the hash isn't part of our canonical view of the chain.

**Description**

This PR adds support for this case by simply looking for the starting
block in the DB and returning it. This stays simple enough and, though
suboptimal, should make the simple case of a peer asking for a sidechain
to work.

It could be further extended by looking for the block's parent through
the parent hash, which should allow us to support cases where `reverse`
is `true`.