Optimize input read
RISC0's official documentation suggests to env::stdin().read_to_end when we need to read the input as raw bytes, because that method doesn't (de)serialization and so do not need to copy or reinterpret the input data. This is the case because we then serialize using rkyv.

On the host side, ExecutorEnvBuilder::write_slice is used to pass in the bytes.