chore(l1): optimize the structure for tracking known txs sent to peers in `TxBroadcaster` (#4675)
**Motivation**

We were seeing spikes of 3~10Gb on post-sync execution, most of it was
due to `resize_rehash` during the `add_txs` of the `TxBroadcaster`.

**Description**

This PR replace the current hashmap of `(tx_hash, peer_id) -> Instant`
for a less duplicated `tx_hash -> (peer_bitmap, last_seen)` + another
`peer_id -> peer_index`. This removes orders of magnitud of entries,
going from 27M to less than 100k. Now all the memory taken from the
TxBroadcaster disappeared from our memory flamegraphs.

WIP: ~I'm right now analyzing the execution of the client throughout the
weekend, we still appear to have a memory issue.~ This was related to a
change in rocksdb that enabled transaction. It didn't affected this fix.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number