feat(replay): endless replay & only eth proofs blocks (#4691)
**Motivation**

We want to remove the `ethrex-replayer` internal tool. This is the first
step; we still need to support requesting multiple RPCs.

**Description**

- Adds the flag `--endless` to `BlockOptions`. This allows replaying
blocks endlessly (once it finishes replaying one block, it fetches the
latest and starts replaying it, and so on).
- Adds the flag `--endless` to `BlocksOptions`. Similar to the above,
but only works if `--from <BLOCK>` is set, replaying endlessly from
`BLOCK`.
- Adds the flag `--only-eth-proofs-blocks` to `BlockOptions` and
`BlocksOptions`. If set, it constrains which blocks are executed.
- Adds the above to the cache commands (the same but skips execution).