chore(replay): remove `ethrex-replayer` bin (#4704)
**Motivation**

This internal tool was useful in a preliminary stage of `ethrex-replay`
and for measuring the prover's performance. Unlike `ethrex-replay`, the
`ethrex-replayer` allowed us to replay L1 blocks endlessly from multiple
RPCs at the same time (Hoodi, Sepolia, Mainnet).

`ethrex-replay` has evolved enough to offer [the "endless" functionality
on its own, under the `blocks`
command](https://github.com/lambdaclass/ethrex/pull/4691). As it doesn't
offer the replay of L1 blocks of multiple RPCs at the same time, this is
the same as running the binary in different processes at the same time
(this is no problem in guest program execution since it's
single-threaded).

This is enough to get rid of `ethrex-replayer`.

**Description**

Removes the `ethrex-replayer` binary.