fix(l2): add script to read address from deployer in docker compose (#4624)
**Motivation**

Running `docker compose up` was broken running on `crates/l2`. The
problem is that the ethrex l2 container reads the `.env` at the moment
of running the `docker compose up` and miss the updates on the file that
writes the deployer.

**Description**

- Deployer now creates the empty `.env` in the shared volume that the
`ethrex_l2` container will read
- Add a parsing and exporting the env vars in the `ethrex_l2` container
as a previous step to running the sequencer.
- Add a copy step on CI jobs to retrieve the addresses used in the
deployment since they are no longer on the host machine
- Change base image of the `Dockerfile` to use a Debian as we need to
use common unix binaries

**How to Test**

On main this should failed (check that you have a fresh start, removing
`cmd/.env`)

```
cd crates/l2
docker compose up
```