feat(l1): implement eth-config (#4625)
**Motivation**

Achieving full Fusaka implementation before launch.

**Description**

This PR implements [EIP-7910](https://eips.ethereum.org/EIPS/eip-7910).
To summarize, it consists of a new eth method, that returns the fork ID,
chain ID, activation time, active precompiles and system contracts for
the currently activated fork, and when applicable, the next and latest
fork's the client has in its configuration.

As part of implementing the precompiles side of things, the logic of how
we check an address is a precompile was refactored, introducing a method
that returns the precompiles per fork so as to more closely tie the
addresses we return in the `eth-config` method and the way we check a
precompile is active when executing it.

To test this PR, you can follow the instructions in [this
page](https://eest.ethereum.org/main/running_tests/execute/eth_config/).
Keep in mind the tests require to have whichever chain they're being run
on fully synced.

Closes #3711