test(l1): full Sync re org tests (#4673)
**Motivation**
We want to test different reorg situations.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
We want to test a reorg where we have 3 nodes. One with the base chain,
one with a side chain and a third one where we want to full sync the
base chain. As per this issue
https://github.com/lambdaclass/ethrex/issues/4617
Adds a test for that scenario, this test is currently flaky, since 50%
of the times it will choose the peer with the base chain, working fine,
but the other 50% it will choose the peer with the side chain, where it
currently fails.
Since there is always a 50% chance of failing, we run the test 30 times
to make sure the probability of failure is small enough.
Other tests are also added:
`test_reorg_back_to_base`: Tests a node that syncs to a side chain and
wants to sync back to the base chain
`test_reorg_back_to_base_with_common_ancestor`: Same test as before but
the base chain has blocks
`test_storage_slots_reorg`: test using a simple contract that sets a
storage slot as received by calldata
`no_reorgs_full_sync_smoke_test`: Tests a normal fullsync without reorgs

Also fixes the reorgs CI check, which is being flaky.

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: Camila Di Ielsi <camila.diielsi@lambdaclass.com>
Co-authored-by: cdiielsi <49721261+cdiielsi@users.noreply.github.com>