perf(l1): improve block headers vec handling in syncer (#4771)
**Motivation**

Currently, in snap and full sync we remove the first header from the vec
because we already have it, this is quite wasteful specially on a Vec
where removing the first element is the most expensive.

**Description**

Converted process_incoming_headers to take an owned iterator instead,
this way we avoid any copy, also in full sync changed the clone and
clear into a single mem take, which both clears and returns the owned
vec to be used as an iterator.

Also avoided some clones when passing block hashes around in store block