feat(l1): snap sync optimizations (#4777)
**Motivation**

This PR reintroduces the optimizations reverted in #4734, now fixed.
There were two problems with it:

- There was a bug in our new (sorted) trie insertion that meant we
incorrectly reconstructed certain storage tries.
- There was a bug in our logic to keep downloading storage leaves on a
pivot jump, where we would not correctly handle accounts whose storage
root went back to a previous value we already had in our trie from a
previous pivot. The way we handle this now is completely best effort: we
try downloading storage leaves for a while and then just switch to
storage healing afterwards. We will improve this when we rewrite snap
sync; the main idea is to differentiate between big storage accounts and
small ones. For big ones we want to ensure we download all their storage
leaves no matter what; for small ones it's fine to do it best effort and
just fallback to storage healing afterwards.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Francisco Xavier Gauna <francisco.gauna@lambdaclass.com>
Co-authored-by: Mario Rugiero <mrugiero@gmail.com>
Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>