feat(l1): add tx broadcasting time interval flag to cli (#4751)
**Motivation**

We've detected that the Request Blob Pooled Transactions Multiple hive
test is failing with:

`FAIL: Error executing step 4: expected 1 hashes, got 5` 

and that's because there's a 1 second interval between broadcasting
transactions that batch several (in this case 5) transactions before
sending them, instead of broadcasting a transaction as soon as it
arrives. To pass the test but still be able to batch transactions (since
it translates in a significative improvement in performance) I'm
introducing a new flag in the cli to set the interval as needed.

**Description**

This pr introduces a new flag in the cli to control the size of the
broadcasting time interval for broadcasting transactions.

This pr depends on this other
[pr](https://github.com/lambdaclass/hive/pull/42) to be merged in
lambda/hive to pass the ci.

Closes #3914