feat(l2): pay base fee to fee vault (#4536)
**Motivation**

We want to optionally configure a `fee vault` address where base fees
are sent instead of being burned.

**Description**

- Modifies our `l2_hook` to send base fees to `fee_vault` if configured.
(In some scenarios, we may still want to burn them (based))
- Adds a new CLI parameter: `block-producer.fee-vault-address`.
- Passes the new parameter to each component that executes transactions:
    - `RpcApiContext`
    - `Committer`
    - `ProofCoordinator`
    - `ProverInput`
- Creates a new endpoint to retrieve the configured `fee_vault` (useful
for `ethrex-replay`).

I found it difficult to decide who should own this value, and how to
avoid passing it throughout the codebase. Any suggestions on this aspect
are welcome!

---------

Co-authored-by: avilagaston9 <gaston.avila@lambdaclass.com>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>