feat(l1): implement flatkeyvalue (snapshots) (#4884)
**Motivation**

We want to have a way to query values directly without going through the
trie (several reads).

To do this, we have a flattened key=>value mapping, which we call
FlatKeyValue. This is sometimes known as "snapshots".

**Description**

A process is launched to generate this mapping in the background. We
ensure `apply_updates` and this process don't conflict, and that block
execution can only make use of the calculated snapshots.

Closes #1997, closes #2409, closes #2410

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: Francisco Xavier Gauna <francisco.gauna@lambdaclass.com>
Co-authored-by: Gianbelinche <39842759+gianbelinche@users.noreply.github.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Mario Rugiero <mrugiero@gmail.com>
Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>
Co-authored-by: fmoletta <99273364+fmoletta@users.noreply.github.com>
Co-authored-by: cdiielsi <49721261+cdiielsi@users.noreply.github.com>
Co-authored-by: Edgar <git@edgl.dev>
Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>
Co-authored-by: Leandro Ferrigno <leanrafa@gmail.com>