refactor(l1,l2): release workflow (#4986)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We are deploying releases without the ability to test them before.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
The new workflow is:
- All new versions should be tagged as RC first (e.g., `v6.0.0-rc.1`).
- The CI will compile binaries and docker images, and will create a
pre-release in Github.
- New RC can be made and the process will repeat.
- Once the binaries were confirmed to work, the pre-release needs to be
edited, changing the name and creating a new tag without the
"subversion" (e.g., `v6.0.0`). Also, the release should be set to
`latest` now.
- The CI will upload the previous compiled docker images with new tags
`latest` and final version.

Changelog was improved too. Now it takes commits since the first common
parent of the previous version (not RC)

<!-- Link to issues: Resolves #111, Resolves #222 -->