fix(l1): fix big account download stopping prematurely (#4994)
**Motivation**

There was a bug that was prematurely marking big accounts as done,
causing inconsistent behaviour in the downloads of these accounts. This
PR aims to fix the bug

**Description**

- When we receive a "should continue" in `verify_range` the task result
will include the account as done in the `remaining_start` field. We
changed the logic of marking accounts done to check that no intervals
remain.

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>