fix(l1): retry on no peers on update pivot (#5002)
**Motivation**

During update pivot, if the `get_best_peer` returned `None` we would
exit the snap sync process, as without peers we cannot progress. There
was an issue when that function returned `None` because the multiplex
request limits were filled, as the peers will eventually timeout the
pending requests and the function would return.

**Description**

- Changed `update_pivot` function to retry on `get_best_peer` not found
- Added sleep to give time to requests to timeout
- Added debug log warning if there is no peers
- Removed superfluous `NoPeers` error