perf(l1, l2): move post-execution trie updates to the background (#4989)
**Motivation**

This PR moves the trie portion of the `apply_updates` function to a
background thread, improving block execution times. This background
thread receives messages through a channel to apply new trie updates and
does two things:

- First, it updates the in-memory diff layers and notifies the process
that sent the message (i.e. the block production thread) so it can
continue with block execution.
- Second, it performs the logic of persisting the bottom-most diff layer
to disk. To do this, we first pause the ongoing thread in charge of
generating the snapshots (i.e. FlatKeyValue) through a message, then
persist the diff layer to disk, then notify again the thread to continue
snapshot generation.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>