refactor(l1,l2): simplify Store::chain_config
Essentially, this condenses two things:
- The fact ChainConfig is only inialized once, while there is a single
  owner of `Store`;
- The fact any `get_chain_config` that fails implies a fatal failure.

From the first point, we infer we can make `set_chain_config` operate on
`&mut Store` rather than use interior mutability.
From the second point, we can make `get_chain_config` infallible, as the
only thing preventing us from doing so was the lock around the field.
This latter change was transitive to most ends, with the exception of
`VmDatabase`: because the `WitnessDB` has a lock around all its members,
it was deemed inappropriate to remove the fallibility. We might revisit
this decision later, as failure to take that lock should probably be
fatal.