fix(l2): verify a single blob proof in the guest program (#5081)
**Motivation**

After Osaka, we (ethrex L2) as blob originators have to update our
approach to constructing blob transactions (see
[this](https://x.com/vdwijden/status/1978443940098457669?s=46&t=UeaOPcmE9dAwr9hxRE1oWQ))
and [include the cell blob proofs in the
bundle](https://github.com/lambdaclass/ethrex/pull/4814).

As a ZK rollup, in the guest program, we verify the blob construction,
but the zkVM libraries lack working functionality for verifying blob
cell proofs (we've created the following issues on SP1 and RISC0,
respectively https://github.com/succinctlabs/kzg-rs/issues/28 and
https://github.com/risc0/risc0/issues/3506).

**Description**

To avoid troubles with libraries, we opted to verify the blob the
old-fashioned way (pre-Osaka-like) in the guest program. This can be
done without any issue, since what's important in the guest program is
to compute the same blob versioned hash as the one that was committed on
the L1; how the blob is verified is independent of this.

This PR reverts the changes done to the prover crate in pos of verifying
cell proofs and updates the prover input generation to generate a single
blob proof for the prover.