feat(l1,l2): support fusaka blobs cell_proofs (#4814)
**Motivation**

We want to support sending blobs with the new specs for Osaka
https://eips.ethereum.org/EIPS/eip-7594. This includes generating
`cell_proofs` for each cell instead a proof for every blob.

**Description**

- The `BlobsBundle::create_from_blobs` method now takes a `Fork`
parameter and generates commitments and proofs differently depending on
whether the fork is Prague or later (e.g., Osaka). This logic is also
reflected in the version field of the bundle.

> [!NOTE]
> Add new `get_l1_active_fork()`. This has a workaround for now since if
no flag `osaka_activation_time` is passed, the L2 will assume it is on
Osaka. If it is passed, it will check the current timestamp if it
activated or not. Created this issue for that
https://github.com/lambdaclass/ethrex/issues/4828
- The EIP-4844 transaction wrapper now exposes a `from_generic_tx`
method that takes a `Fork` and sets the wrapper version and blob bundle
accordingly.

Closes #4779

---------

Co-authored-by: Gianbelinche <39842759+gianbelinche@users.noreply.github.com>
Co-authored-by: Estéfano Bargas <estefano.bargas@fing.edu.uy>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>