fix(l2): remove one-time checkpoint if already exists (#5083)
**Motivation**

In a previous PR, DB checkpoints were introduced to ensure old state
availability in the current path-based fashion. Every time a batch is
sealed, a checkpoint whose state is the state of the latest block of the
sealed batch is created to be used in the next batch.

The checkpoint is needed in two different steps of the batch commitment:
for batch preparation (this is essentially building the batch) and for
witness generation. Both steps need a non-modified checkpoint, but they
both need to modify the checkpoint to be able to re-execute the batch.

As batch preparation occurs before witness generation, we opted to
create a one-time checkpoint out of the main checkpoint that can be
modified during batch preparation if needed (sometimes the batch was
already available in the DB, and there's no need to re-execute
anything); then, witness generation modifies the original checkpoint as
needed because it is no longer needed.

Once the one-time checkpoint fulfills its purpose, it is removed.
Currently, if batch preparation fails, the one-time checkpoint is not
removed, and after retrying batch preparation, there's another attempt
at creating the one-time checkpoint, which ends in an error because the
directory already exists. We need to either avoid creating the one-time
checkpoint again or to remove the existing one.

**Description**

Remove the existing one-time checkpoint if it already exists.

---------

Co-authored-by: avilagaston9 <gaston.avila@lambdaclass.com>
Co-authored-by: Gianbelinche <39842759+gianbelinche@users.noreply.github.com>