fix(l1): don't trigger snap sync if fcu head is already canonical (#4985)
**Motivation**
In hive tests we often receive an fcu with either the genesis or last
imported block as head to notify the node that it is already synced. As
this is not usual in a real-case scenario, we just trigger a snap sync
to the advertised head. This PR solves this by first checking if the fcu
head is already part of our canonic chain before triggering a sync
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* When handling a forkchoice in snap sync mode, check that the head is
not already canonical before triggering a sync, if it is already
canonical, apply the forkchoice
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #4846