ci(l1,l2): automatic workflow trigger to ethrex-apt repository. (#5022)
**Motivation**
Adds a trigger to make the
[lambdaclass/ethrex-apt](https://github.com/lambdaclass/ethrex-apt) repo
trigger it's automatic release workflow procedure when a new release is
made.

**Description**
Added a new job inside the `.github/workflows/tag_release.yaml` file
that triggers the workflow in the APT repo that builds the `.deb` files
for Debian/Ubuntu. This job is only triggered if all the previous steps
finish properly.