fix(l2): save last committed batch as current checkpoint (#5152)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We are saving the latest `Store` as the "current" one. This means the
next time a batch is constructed, it might use a latter checkpoint than
the needed. This can occur if the batch is sealed with less blocks than
the available (e.g., if there's no more available space).

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Save a checkpoint from the batch execution instead

<!-- Link to issues: Resolves #111, Resolves #222 -->