perf(l1): avoid opening tries (#5145)
**Motivation**

Currently to get a storage we are:

* getting the block header to obtain the state root
* getting the account to get the storage root
* reading the value

The first step isn't needed because the state root is the same through
the block, and the second isn't needed because since we're using a
path-based DB we can directly get the value (without even using the
trie, if the FKVs are computed).

**Description**

We make the VM database ask by state_root instead of block hash
(skipping step 1), and we skip reading the account storage root when the
FlatKeyValues are computed for it.

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>