perf(l1): validate receipts in the same thread as the execution (#5170)
**Motivation**

In pipelined execution we have two threads: executor and merkelizer.
Currently we perform validations (gas usage, receipts, requests) after
both are done. However, these do not depend on the results of the
merkelization so they can be merged with the execution thread.

**Description**

Since the merkelization thread necessarily finishes after execution is
complete, we don't need to worry about the executor taking too long.

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>