Avoid unnecesary work on bloom and keep a per layer bloom
Keeping a per layer bloom helps avoiding rehashing every time a bloom
rebuild happens.
This is because the bloom merge method uses the u64 hash keys directly
instead of rehashing the value of type T.

This is an improved version of https://github.com/lambdaclass/ethrex/pull/5175
where in addition to that we keep a per layer bloom to avoid extra hash
work.