feat(l1): admin_addPeer endpoint (#5004)
**Motivation**

Include geth's
[admin_addPeer](https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-admin#admin-addpeer)
endpoint since we found it necessary for running eip7934 hive tests with
the `ethereum/eels/consume-sync ` simulation.
Besides Nethermind, Reth, Besu, Erigon and other execution clients
implement it.

**Description**

This pr introduces the RPC endpoint `admin_addPeer`. Since this endpoint
needed to access p2p's generic server used to establish connections, the
`init_l1` function was restructure so the genserver is saved within the
peer_handler struct and can later be accessed from the `RPCApiContext`.

In `admin_addPeer` the connection is tried in a period of 10 seconds
waiting 100 milliseconds in between each message casted. This is because
connections are asynchronous so to check if the connection with the peer
was made we need to wait until it's established.

---------

Co-authored-by: SDartayet <sofiadartayet@gmail.com>
Co-authored-by: SDartayet <44068466+SDartayet@users.noreply.github.com>
Co-authored-by: ElFantasma <estebandh@gmail.com>