fix(l1): reset `ctr` variable to 0 once committed (#5214)
**Motivation**

There seems to be a logic error in the flat-key-value generator on the
store. A counter wasn't being correctly reset once it reached a
threshold, effectively making us commit on each item inserted.

**Description**

This PR resets the counter on each commit. This should make batching
work correctly from now on.