fix(l2): peer table duplication (#5227)
**Motivation**

PR #5004 led to changes in `init_l2()` that weren't properly updated, so
now when initializing L2 there are 2 Peer Tables.

**Description**

This pr uses the Peer Table created in the first place to initialize the
Peer Handler instead of creating a new one.
It also updates the gen server initialization for the gen server dummy.