refactor(l1): bring all test modules in the rpc crate into a single test_util module (#5213)
**Motivation**

Stop having all the dummy() functions that create non-working components
spread all around the code.

**Description**

This pr brings all test utils in the rpc crate into a single test_utils
module, including all the dummy() functions that were spread all around
the code.

This pr only focuses in the rpc crate to tackle issue #5070 since it's
the only one that had a test_utils module and required having a dummy
instance of a component to be created for testing.

Closes #5070