fix(l1,l2): fix occasional error with `debug_executionWitness` (#5212)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Sometimes when we want a witness ethrex returns an error, this aims to
fix it.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
1. There's a bug that assumes that we have some data in a map that we
don't necessarily have. This is usually triggered when there's a call to
the BLOCKHASH opcode during a block. The idea is to requests all the
headers of intermediate blocks so that we can generate the witness
correctly.
2. Now the Guest Program accepts blocks with hash already computed, it
just recomputes it and compares the hash with the already set one (if
any), returning an error in case they differ. This was suggested in
https://github.com/eth-act/zkevm-benchmark-workload/issues/177 and we
liked the idea.

Both changes are in the same PR because the goal was just to solve `1`
but as I was hashing the blocks when generating the witness I decided to
fix `2` elegantly so we can pass the Blockchain tests.

Closes #5250

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>