perf(l1): batch merkelization requests if the queue is not empty (#5200)
**Motivation**

Currently for every transaction being executed, we get the transitions
and send the updates to the merkelization thread.

However this has two drawbacks:
- getting the update list isn't cheap, taking up 10% of block execution
- unbatched requests cause more work to be done, increasing total
merkelization time

The merkelization queue often had dozens of requests by the end of block
execution.

**Description**

We delay calculating and sending updates until the merkelizer is ready
(queue is empty), which causes the changes to the batched.

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>