fix(l1): sleep between FCU and getPayload calls in dev-mode block producer (#5310)
**Motivation**

Our `--dev` block production had a bug where we waited an interval
before sending an FCU and called getPayload immediately after, causing
inclusion latency to be 1.5x the interval.

**Description**

This PR moves the sleep between the FCU and getPayload calls, fixing the
issue. It also switches to `tokio::time::interval` for more precision.