fix(l2): checkpoint creation (#5321)
**Motivation**

This PR addresses two main issues:

- Sealing a batch, storing the `prover inputs`, and creating the
checkpoint for that batch are currently performed non-atomically. If the
node restarts during any of these three operations, the committer will
end up in an invalid state.
- The batch checkpoint is created at the beginning of the building
process. If an error occurs while building the batch, the `l1_committer`
gets stuck with the following error:

```
2025-11-12T15:47:26.129936Z ERROR L1 Committer Error: Committer failed retrieve block from storage: Failed to create RocksDB checkpoint at "dev_ethrex_l2/checkpoint_batch_1": Invalid argument: Directory exists
```

**Description**

- Stores the batch and `prover inputs` atomically in the rollup storage.
- When the `l1_committer` encounters a batch generated in a previous
iteration, it now checks whether the corresponding checkpoint exists. If
it doesn't, the committer creates it by re-executing the batch.


Closes None