perf(l1,l2): avoid unnnecessary code hashing
Implements the `Code::from_bytecode_unchecked` function to create the
code object from a trusted or bogus hash.

The trusted case is used for snap sync, where the downloader task
already checks the hash is correct, so it can pass it down.

The bogus hash case is for init code, where the hash is never used
(there is no account pointing to it, no entries in the DB addressed by
it and no opcode can access the current code's hash, only the hash for
an account associated to an address).