fix(l1,l2): fix encoding of nodes in witness (#5421)
**Motivation**

- Fix `debug_executionWitness` endpoint

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- When asking for the witness of a block the trie nodes were being
encoded 2 times (double encoding). This just removes the additional
encoding.

This has been tested with ethrex-replay and now it works perfectly fine
with ethrex nodes. We were getting errors decoding RLP before.


How to test it?
- Run L1 in dev mode
- Request witness with debug_executionWitness

See the response both in main and in this branch, you'll see that the
nodes are "the same" but in main they are encoded one more time.
Basically in main: `RLP(RLP(node))`, but here: `RLP(node)`. We want the
latter.