fix(l2): fix ZisK install (#5430)
**Motivation**

Sometimes GitHub runners are slow to download from Google Cloud Storage,
which is the case for the ZisK toolchain install. This downloads a
proving key (big file) from GCS. This file is actually not necessary
unless we want to generate a proof in a CI workflow.

Setting the `SETUP_KEY` env var should skip downloading this file.