refactor(l1): refactor chainconfig (#5233)
Motivation

Improve the maintainability of our code with regards to future forks.

Description

Currently we have a lot of match statements spread throughout various
functions to get the current fork, blobschedule, etc; and separate
functions for checking the activation of each fork. This creates serious
maintainability issues down the line. This PR centralized the match
statements in two functions: one to get the activation timestamp/block
for a fork, and one to get the blob schedule for a fork, and rewrites
the rest of the functions accordingly. All the funcitons to check fork
activation were also aggregated into a single `is_fork_activated`
function that takes the fork as a parameter and checks its activation.
The implementation of `get_blob_schedule_for_time` also fetches the most
recent active blob schedule if the current fork specifies none, which is
part of tackling #4849.

Additionally, some code related to pre-Merge logic was removed (checks
for EIP 155 and EIP 2028)

Closes #4720

---------

Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>