fix(l1): reduce spammy trace-level logs during snap-sync (#5492)
**Motivation**

Running ethrex with trace-level logs results in a lot of spam of "We
didn't get a peer from the table" or similar, due to busy loops we have
when we don't have many peers. This is especially important for the
first few stages, since we start with a small number of peers.

**Description**

This PR reduces the spam by adding a sleep of 10ms, plus a counter to
make the warning be printed only once every 1000 times (effectively once
every 10s). Also does the same in other stages of snap-sync, adding logs
in other busy loops to detect if/when we land in those cases instead of
silently ignoring them.