feat(l1): pin `execution-apis` version when running `rpc-compat` hive suite in Makefile targets (#5463)
In the hive workflow that runs for each PR, we set the execution-apis
version as a building when running `rpc-compat` testing suite, as the
latest version uses a pre-merge genesis file. This causes hive tests to
pass on the CI but not when developing locally. This PR fixes this by
adding the buildarg flag when running rpc-compat testing suite