chore(l1): remove hardcoded platform for ethrex in docker compose (#5594)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We don't need to specify the `platform` for the docker image anymore, as
we now build it for both `amd64` and `arm64`. Also, `amd64` images are
not working anymore on arm machines due to AVX.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Remove the platform specified in the root's docker compose so it
auto-detects which to use.

<!-- Link to issues: Resolves #111, Resolves #222 -->