fix(l1): ignore FCUs if state is not reachable (#5563)
**Motivation**

We currently fail to sync if we receive an FCU for a block we can't
reach from our DB.

**Description**

This PR ignores FCUs when the nearest link to the chain is a block with
a state we pruned, or before our last snap-sync pivot. It also logs a
warning, which recommends that the user re-sync their node if they
consistently see it. Note that this is a temporary patch, which should
be removed once we rewrite snap-sync.

Closes #5558

---------

Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>