fix(l1): validate block body on import (#5606)
**Motivation**

We don't currently validate that header and body match when importing
blocks. This could cause unclear errors when working on invalid data, so
we should check that. Also, some hive tests were failing because of
that.

**Description**

This PR adds to `import_blocks` and `import_blocks_bench` a check that
header and body match.

Closes #5601

Related to #5605