fix(l1): error on missing parent (#5559)
**Motivation**

Currently if the state trie root does not match with the one being used
for execution, we fail due to various errors that happen as a result of
executing from incorrect state (invalid nonces, incorrect gas, state
root mismatch) instead of alerting about the real error.

While this should only happen due to a bug, it has happened several
times.

**Description**

Errors if the state root being used for execution isn't found.

**Checklist**

- [X] Updated `STORE_SCHEMA_VERSION` (crates/storage/lib.rs) if the PR
includes breaking changes to the `Store` requiring a re-sync.

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>