fix(l2): fix block execution with SP1 (#5542)
**Motivation**

Execution with SP1 was failing after PR
https://github.com/lambdaclass/ethrex/pull/5529 which changes a line of
the `ecmul` precompile to make it compatible with an API difference
between the substrate-bn patch of SP1 and ZisK.

Apparently this also changes the behavior of the multiplication for the
SP1 patch, and some blocks (like block 23919500 from Mainnet) were
failing to execute due to a gas mismatch error.

Reverting the change for SP1 fixes the issue.

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>