fix(l1): continue on PeerTable call timeout (#5642)
**Motivation**

We're having an error where snap-sync exits due to an internal timeout
in a call to `PeerTable`. We're still looking into why this timeout
happens, but when this happens, we should simply retry.

**Description**

This PR copies the `inspect_err().unwrap_or(None)` pattern from
`request_account_range` to `request_storage_ranges` and
`request_bytecodes`.

Closes #5637