feat(l1): add discv5 session structures and remaining official vector tests (#5673)
**Motivation**

Adds a Session struct and related functions to the discv5 module, this
represents a session established from a handshake. This PR does not
implement the handshake protocol, but it lays a path to doing so easily.

This is based on reading
https://github.com/ethereum/devp2p/blob/master/discv5/discv5-theory.md#sessions

This allows us to add the remaining test vectors related to
cryptography, which is done now.

Also renamed the PacketDecodeErr to PacketCodecError, since the error
was also used when encoding and for other things.

Fixed nonce generation to be secure according to the spec