feat(l1): loki logs integration and dashboard changes (#5472)
**Motivation**

Adding logs to grafana as part of the local monitoring stack

**Description**

This PR add loki and promtail to our current metics docker compose. To
maintain compatibility with how we are currently running we made a
couple of decisions:
- Continue logging on stdout by default
- Add a new `--log.dir` cli param which enables logging files (alongside
the stdout)
- Use `./logs` by default in the repo as the folder from which take logs
(and make them configurable by an env variable)
- Add a new Logs row (collapsed by default) to our dashboard.
- Add documentation and update the dashboard docs

<img width="2248" height="1030" alt="image"
src="https://github.com/user-attachments/assets/19c6849f-9aea-4329-a5f8-3b2b171c731a"
/>

**Next Steps**
There are a couple of things to improve:
- This set up is working for local execution but not for centralized
ones, we mnight need some changes to how loki and promtail interacts for
that
- The docker-compose now continue to works as default (`STDOUT`) we
might want to move to add logs and show them on dashboards by default in
the future. That might also require to launch the metric related
services.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3265