feat(l1): discv5 (#5655)
**Motivation**

Ethereum nodes are expected to run `discv5` for early 2026. 

**Description**

On one hand, this is work in progress, but in the other we don't want to
have a huge PR at the end of development. So we plan to start merging
`discv5` work into main, but behind a feature flag so nothing is
compiled or run, except for development. The flag is already present in
this commit, so we should be clear to merge.

**Checklist**

- [ ] Updated `STORE_SCHEMA_VERSION` (crates/storage/lib.rs) if the PR
includes breaking changes to the `Store` requiring a re-sync.

Closes #4491

---------

Co-authored-by: MrAzteca <azteca1998@users.noreply.github.com>
Co-authored-by: Edgar <git@edgl.dev>