feat(l1): exposing node version as a metric (#5692)
**Motivation**

Expose ethrex version as a metric to avoid depending on
ethereum-metrics-exporter for it.

**Description**

This PR adds `ethrex_info` as a new metric that expose version, build
related information, and network.

```
# HELP ethrex_info Node information including version and build details
# TYPE ethrex_info gauge
ethrex_info{branch="expose-ethrex-version",commit="5e88113",network="hoodi",rust_version="1.90.0",target="aarch64-apple-darwin",version="7.0.0"} 1
```

This allowed us to stop relying on `ethereum-metrics-exporter` on our L1
dashboard (except for the fork one which we'll need a different approach
and a follow-up PR), the dashboard and docs are updated.

<img width="2548" height="300" alt="image"
src="https://github.com/user-attachments/assets/2348c46f-0a65-44e5-bb38-6d5de59fe803"
/>

There is also a small change in how we treat gather metric error for the
first metrics configured to be consistent with what we are doing today.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Resolves #5549