chore(l2): bump ZisK to 0.15.0 & add modexp syscall (#5694)
0.15.0 introduces a new modexp syscall and fixes for the existing
patches, apart from other optimizations (see
https://github.com/0xPolygonHermez/zisk/releases/tag/v0.15.0)

we found a regression some time ago in main, probably related to a ZisK
patch, in which some blocks failed to execute with a gas mismatch error.
We tested those blocks again and the regression was gone. So the new
patches' versions contain fixes too.