fix(l2): revert substrate-bn patched `ecadd` precompile for SP1 (#5723)
**Motivation**

https://github.com/lambdaclass/ethrex/pull/5535 introduced a version of
ecadd that uses substrate-bn, because both ZisK and SP1 patch this crate
to use zkVM accelerators for the operation.

After merging, we found an Ethereum Mainnet block failing to execute
with a gas mismatch error, which does not happen one commit before
merging that PR.

This PR reverts the usage of substrate-bn for SP1 only.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

**Checklist**

- [ ] Updated `STORE_SCHEMA_VERSION` (crates/storage/lib.rs) if the PR
includes breaking changes to the `Store` requiring a re-sync.

Closes #issue_number